# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "InvoiceTool Swagger API",
        "description": "InvoiceTool Swagger API description",
        "contact": {
            "email": "admin@finpay.com"
        },
        "version": "1.0.0"
    },
    "host": "invoice-tool-nov8.test/",
    "basePath": "api/v1",
    "paths": {
...
        "/Taxrate": {
            "get": {
                "tags": [
                    "SETTINGS - TAX RATES"
                ],
                "summary": "Taxrate",
                "operationId": "bid",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SETTINGS - TAX RATES"
                ],
                "summary": "Tax Store",
                "description": "Tax Store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "TaxRate Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "compound",
                        "in": "formData",
                        "description": "Compound",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "percentage",
                        "in": "formData",
                        "description": "Tax Percentage",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "calculate_as_vat_gst",
                        "in": "formData",
                        "description": "Calculate as VAT/GST",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/Taxrate/{id}": {
            "get": {
                "tags": [
                    "SETTINGS - TAX RATES"
                ],
                "summary": "Tax Edit",
                "operationId": "Teid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tax Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SETTINGS - TAX RATES"
                ],
                "summary": "Tax Update",
                "description": "Tax Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tax Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "TaxRate Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "compound",
                        "in": "formData",
                        "description": "Compound",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "percentage",
                        "in": "formData",
                        "description": "Tax Percentage",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "calculate_as_vat_gst",
                        "in": "formData",
                        "description": "Calculate as VAT/GST",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SETTINGS - TAX RATES"
                ],
                "summary": "Tax Delete",
                "operationId": "Tdid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tax Delete",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}